refactor: move attribute classes out of the section folder
