+++ src/Donations/Endpoints/DonationUpdateAttributes/StatusAttribute.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Give\Donations\Endpoints\DonationUpdateAttributes\PaymentInformation;
+namespace Give\Donations\Endpoints\DonationUpdateAttributes;
 
-use Give\Donations\Endpoints\DonationUpdateAttributes\DonationUpdateAttribute;
 use Give\Donations\Models\Donation;
 use Give\Donations\ValueObjects\DonationStatus;
